Added train stops and update them

+++ src/main/java/com/trs/trs_admin_service/repo/TrainRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TrainRepository extends JpaRepository<Train, Long> {
+public interface TrainRepository extends JpaRepository<Train, Integer> {
     Optional<Train> findByTrainNumber(Integer trainNumber);
 }
